uapi: convert drivers/tty.h
